#9 fixed money transferring

+++ src/database/bets.py
@@ -2,7 +2,7 @@ from src.mongodb import bet_col
 from models.bet import Bet
 
 
-async def put_new_bet(bet_value: int, end_time: int, bot_id: int, game_id: str):
+async def put_new_bet(bet_value: float, end_time: int, bot_id: int, game_id: str):
     bet = Bet(
         id=await bet_col.count_documents({}),
         bet_value=bet_value,
